Fix utils/storage extension from svelte to ts
